Fixed upload of large packages
In case a large file is uploaded the pcre.backtrack_limit might
not be high enough to get all the file contents via preg_match.
To circumvent this problem we match against a substring of the
request and capture the offsets (PREG_OFFSET_CAPTURE) so we can
later use the captured offset and substr to get the file's contents.